move cli_entrypoint to nemo/lightning/run/cli.py
