+++ nemo/collections/llm/api.py
@@ -26,6 +26,7 @@ import nemo.lightning as nl
 from nemo.lightning import AutoResume, NeMoLogger, OptimizerModule, Trainer, io
 from nemo.lightning.pytorch.callbacks import PEFT, ModelTransform
 from nemo.utils import logging
+from nemo.lightning.run import cli_entrypoint
 
 if TYPE_CHECKING:
     from megatron.core.inference.common_inference_params import CommonInferenceParams
@@ -33,22 +34,6 @@ if TYPE_CHECKING:
 
 
 TokenizerType = Any
-
-
-def cli_entrypoint(*args0, **kwargs0):
-    def decorator(func):
-        def wrapper(*args, **kwargs):
-            try:
-                import nemo_run as run
-
-                return run.cli.entrypoint(*args0, **kwargs0)(func, *args, **kwargs)
-            except (ImportError, ModuleNotFoundError) as error:
-                logging.warning(f"Failed to import nemo.collections.llm.[api,recipes]: {error}")
-                return func(*args, **kwargs)
-
-        return wrapper
-
-    return decorator
 
 
 @cli_entrypoint(namespace="llm")
